+++ builders/php.js
@@ -6,32 +6,25 @@ const path = require('path');
 const semver = require('semver');
 const addBuildStep = require('./../utils/add-build-step');
 
-const composerVersions = {
-  '1': '1.10.27',
-  '2': '2.8.3',
-  '2.2': '2.2.24',
-};
-
 /**
  * Get the appropriate Composer version based on the PHP version.
- * @param {string} phpVersion - The PHP version.
+ * @param {semver} phpSemver - The PHP semantic version.
  * @return {string|boolean} - The Composer version or false if we cannot parse the version.
  */
-const getDefaultComposerVersion = phpVersion => {
-  phpVersion = semver.coerce(phpVersion);
+const getDefaultComposerVersion = phpSemver => {
   // Don't set a default composer version if we cannot
   // parse the version such as with `custom`.
-  if (!phpVersion) return false;
+  if (!phpSemver) return false;
 
-  if (semver.lt(phpVersion, '5.3.2')) {
+  if (semver.lt(phpSemver, '5.3.2')) {
     // Use Composer 1 for PHP < 5.3.2
-    return composerVersions['1'];
-  } else if (semver.lt(phpVersion, '7.3.0')) {
+    return '1';
+  } else if (semver.lt(phpSemver, '7.3.0')) {
     // Use Composer 2.2 LTS for PHP < 7.3
-    return composerVersions['2.2'];
+    return '2.2';
   } else {
     // Use Composer 2 for PHP >= 7.3
-    return composerVersions['2'];
+    return '2';
   }
 };
 
@@ -60,13 +53,25 @@ const nginxConfig = options => ({
 const xdebugConfig = host => ([
   `client_host=${host}`,
   'discover_client_host=1',
- 'log=/tmp/xdebug.log',
- 'remote_enable=true',
+  'log=/tmp/xdebug.log',
+  'remote_enable=true',
   `remote_host=${host}`,
 ].join(' '));
 
-/*
- * Helper to build a package string
+/**
+ * Helper function to build a package string by combining package name and version
+ *
+ * @param {string} pkg - The package name
+ * @param {string} version - The package version
+ * @return {string} The formatted package string, either "pkg:version" or just "pkg" if version is empty
+ *
+ * @example
+ * // Returns "php:7.4"
+ * pkger('php', '7.4');
+ *
+ * @example
+ * // Returns "mysql"
+ * pkger('mysql', '');
  */
 const pkger = (pkg, version) => (!_.isEmpty(version)) ? `${pkg}:${version}` : pkg;
 
@@ -172,11 +177,15 @@ module.exports = {
       // Merge the user config onto the default options
       options = parseConfig(_.merge({}, config, options));
 
+      // Get the semver of the PHP version, NULL if we cannot parse it
+      const phpSemver = semver.coerce(options.version);
+      phpSemver && debug('Parsed PHP semantic version: %s', phpSemver);
+
       // Mount our default php config
       options.volumes.push(`${options.confDest}/${options.defaultFiles._php}:${options.remoteFiles._php}`);
       options.volumes.push(`${options.confDest}/${options.defaultFiles.pool}:${options.remoteFiles.pool}`);
       // Shift on the docker entrypoint if this is a more recent version
-      if (options.version !== 'custom' && semver.gt(semver.coerce(options.version), '5.5.0')) {
+      if (phpSemver && semver.gt(phpSemver, '5.5.0')) {
         options.command.unshift('docker-php-entrypoint');
       }
 
@@ -202,28 +211,38 @@ module.exports = {
       };
       options.info = {via: options.via};
 
-      // Add our composer things to run step
+      // Determine the appropriate composer version to install if not specified
+      if (options.composer_version === true || options.composer_version === '') {
+        options.composer_version = getDefaultComposerVersion(phpSemver);
+      } else if (typeof options.composer_version === 'number') {
+        options.composer_version = options.composer_version.toString();
+      }
+      const usingComposer1 = options.composer_version && semver.satisfies(options.composer_version, '1.x');
+
+      // Add prestissimo as a global package for Composer 1.x performance improvements. Requires PHP >= 5.3
+      if (usingComposer1 && phpSemver && semver.gte(phpSemver, '5.3.0')) {
+        options.composer = options.composer || {};
+        options.composer = {'hirak/prestissimo': '*', ...options.composer};
+      }
+
+      // Add build step to enable xdebug
+      if (options.xdebug) {
+        addBuildStep(['docker-php-ext-enable xdebug'], options._app, options.name, 'build_as_root_internal');
+      }
+
+      // Add build step to install our Composer global packages
       if (!_.isEmpty(options.composer)) {
         const commands =
           require('../utils/get-install-commands')(options.composer, pkger, ['composer', 'global', 'require', '-n']);
         addBuildStep(commands, options._app, options.name, 'build_internal');
       }
 
-      // Add activate steps for xdebug
-      if (options.xdebug) {
-        addBuildStep(['docker-php-ext-enable xdebug'], options._app, options.name, 'build_as_root_internal');
-      }
-
-      // Determine the appropriate composer version if not already set
-      if (options.composer_version === true || options.composer_version === '') {
-        options.composer_version = getDefaultComposerVersion(options.version);
-      }
-
-      // Install the desired composer version
+      // Install the desired composer version as the first `build_internal` build step
       if (options.composer_version) {
         debug('Installing composer version %s', options.composer_version);
         const commands = [`/helpers/install-composer.sh ${options.composer_version}`];
-        addBuildStep(commands, options._app, options.name, 'build_internal', true);
+        const firstStep = true;
+        addBuildStep(commands, options._app, options.name, 'build_internal', firstStep);
       }
 
       // Add in nginx if we need to